DT467: Add GetPluginPath and RegistryService
GetPluginPath returns the full path of the
ParatextPluginManager. This currently only works
assuming that Paratext8 has been installed and
that the system is 64-bit.

RegistryService handles reading from the registry.
My plan is to use this to mock registry calls so
that we can test the logic that determines which
subkey to look for in the registry.