DT467: Scan for plugins by checking the PT plugins directory
Rather than maintaining a list of installed plugins, the plugin
manager will now scan all sub-directories in the appropriate
directory, looking for JSON files. If a JSON file is found, it will
be treated as a PluginDescription.

If we expect plugins to contain other JSON files, we will need
to handle that differently. The code currently expects only
a PluginDescription and will throw an error if certain fields are
missing from the file.